Updated imports in dynamics modules and visualization example to use newest topology functions
